+++ utils/sample_group_aligned_code.cpp
@@ -47,7 +47,7 @@ int main(int argc, char** argv)
     {
       std::vector<char>::iterator biter = buffer.begin();
       std::vector<char>::iterator hiter = buffer.begin();
-      for (int j = 0; j < num_sample; ++ j) {
+      for (size_t j = 0; j < num_sample; ++ j) {
 	const uint32_t value = random();
 	
 	const size_t offset = utils::group_aligned_encode(value, &(*hiter), j);
